Colision detection big progress, still some bugs when falling next to the wall too close, player can walk in to the collision object in some cases

+++ include/gravitation.h
@@ -33,7 +33,7 @@ public:
 	gravitation(textureStateManager & stateObject);
 	~gravitation()=default;
 	
-	void antigravityForce(int strength);
+	void antigravityForce(float strength);
 	gravityDetails&  getHeight(gravityDetails& currentVerticalVelocity);
 
 	inline float gravityEqeuation(int x) {
